+++ app.js
@@ -1,3 +1,30 @@
+class Registration {
+    constructor(bot) {
+        this.bot = bot;
+        bot.on(['/start', '/hello'], this.registrationFunc);
+
+        bot.on('newChatMembers', this.registrationFunc);
+
+        bot.on(['/gaf_team'], (msg) => {
+            let person = registerationMap.get(msg.from.id);
+            if (person !== undefined) {
+                person.gaf = msg.text.split(' ')[1];
+                person.team = msg.text.split(' ')[2];
+                // registerationMap.set(msg.from.id, person);
+                msg.reply.text("welcom " + person.name + " from " + person.gaf + "/" + person.team);
+            }
+        });
+    }
+
+    registrationFunc(msg) {
+        msg.reply.text('הכנס גף וצוות /gaf_team')
+        let person = {};
+        person.id = msg.from.id;
+        person.name = msg.from.first_name + " " + msg.from.last_name;
+        registerationMap.set(person.id, person);
+    };
+}
+
 const TeleBot = require('telebot');
 const bot = new TeleBot('605816885:AAFe49RlVOiOQjU0F5Uqd60PorjSwJ2klmk');
 
@@ -32,7 +59,7 @@ let replyMarkup = bot.keyboard([
     [replayOptions.VACATION.name, replayOptions.COURSE.name]
 ], {resize: true});
 
-
+new Registration(bot);
 bot.on('/start', msg => {
     _isWaitingForAnswer = true;
 return bot.sendMessage(msg.from.id, 'בוקר טוב, איפה אתה?', {replyMarkup});
