+++ bin/www
@@ -4,7 +4,7 @@
  * Module dependencies.
  */
 
-var app = require('../app');
+var app = require('../registration');
 var debug = require('debug')('untitled2:server');
 var http = require('http');
